+++ tezt/manual_tests/baker_test.ml
@@ -56,8 +56,8 @@ type status = {
   validated : (string * bool) option;
   applied : string option;
   expected_consensus_ops : int;
-  endorsements : String_set.t;
-  preendorsements : String_set.t;
+  attestations : String_set.t;
+  preattestations : String_set.t;
 }
 
 let init_status expected_consensus_ops =
@@ -65,8 +65,8 @@ let init_status expected_consensus_ops =
     validated = None;
     applied = None;
     expected_consensus_ops;
-    endorsements = String_set.empty;
-    preendorsements = String_set.empty;
+    attestations = String_set.empty;
+    preattestations = String_set.empty;
   }
 
 (* This function only returns when all operations have been processed by the
@@ -102,15 +102,15 @@ let check_ophs_in_mempool ~applied ophs (mempool : Mempool.t) =
          "Not all expected operations are classified as %s in the mempool"
          (if applied then "applied" else "branch_delayed"))
 
-(* This test checks that the baker can start to pre-endorse a proposal before
-   the block application finishes in the node. On Mumbai, pre-endorsement should
+(* This test checks that the baker can start to pre-attest a proposal before
+   the block application finishes in the node. On Mumbai, pre-attestation should
    not be propagated and PQC should not be reached before the block
    application. *)
-let baker_early_preendorsement_test =
+let baker_early_preattestation_test =
   Protocol.register_test
     ~__FILE__
-    ~title:"Test baker early pre-endorsement"
-    ~tags:["node"; "baker"; "early"; "preendorsement"]
+    ~title:"Test baker early pre-attestation"
+    ~tags:["node"; "baker"; "early"; "preattestation"]
     ~supports:(Protocol.From_protocol 16)
   @@ fun protocol ->
   Log.info
@@ -224,12 +224,12 @@ let baker_early_preendorsement_test =
     in
     return
     @@ List.filter_map
-         (fun endorsing_right ->
-           let round = JSON.(endorsing_right |-> "round" |> as_int) in
+         (fun attestation_right ->
+           let round = JSON.(attestation_right |-> "round" |> as_int) in
            if round > max_round then None
            else
              let key =
-               JSON.(endorsing_right |-> "consensus_key" |> as_string)
+               JSON.(attestation_right |-> "consensus_key" |> as_string)
              in
              List.find_map
                (fun (bootstrap : Account.key) ->
@@ -352,7 +352,7 @@ let baker_early_preendorsement_test =
   let* () = baker_t in
 
   let baker_on_event ?(log = false) (status : status ref)
-      preendorsements_resolver pqc_resolver endorsements_resolver
+      preattestations_resolver pqc_resolver attestations_resolver
       Baker.{name; value; timestamp = _} =
     match name with
     | "new_valid_proposal.v0" ->
@@ -367,47 +367,47 @@ let baker_early_preendorsement_test =
             Option.map (fun v -> (fst v, true)) !status.validated
           in
           status := {!status with validated}
-    | "preendorsement_injected.v0" -> (
+    | "preattestation_injected.v0" | "preendorsement_injected.v0" -> (
         match (!status.validated, !status.applied) with
         | Some (_, false), Some _ ->
             failwith
-              "Pre-endorsement should be injected before application of the \
+              "Pre-attestation should be injected before application of the \
                block"
         | Some (_, false), None ->
             status :=
               {
                 !status with
-                preendorsements =
+                preattestations =
                   String_set.add
                     JSON.(value |-> "ophash" |> as_string)
-                    !status.preendorsements;
+                    !status.preattestations;
               } ;
             if
-              String_set.cardinal !status.preendorsements
+              String_set.cardinal !status.preattestations
               = !status.expected_consensus_ops
-            then Lwt.wakeup preendorsements_resolver ()
+            then Lwt.wakeup preattestations_resolver ()
         | None, _ | Some (_, true), _ ->
-            () (* Only pre-endorsement for the wanted proposal are recorded *))
-    | "endorsement_injected.v0" -> (
+            () (* Only pre-attestation for the wanted proposal are recorded *))
+    | "attestation_injected.v0" | "endorsement_injected.v0" -> (
         match (!status.validated, !status.applied) with
         | Some (_, false), None ->
             failwith
-              "Endorsement can only be injected once the block is applied"
+              "Attestation can only be injected once the block is applied"
         | Some (b1, false), Some b2 when String.equal b1 b2 ->
             status :=
               {
                 !status with
-                endorsements =
+                attestations =
                   String_set.add
                     JSON.(value |-> "ophash" |> as_string)
-                    !status.endorsements;
+                    !status.attestations;
               } ;
             if
-              String_set.cardinal !status.endorsements
+              String_set.cardinal !status.attestations
               = !status.expected_consensus_ops
-            then Lwt.wakeup endorsements_resolver ()
+            then Lwt.wakeup attestations_resolver ()
         | None, _ | Some (_, true), _ | Some (_, false), Some _ ->
-            () (* Only endorsement for the wanted proposal are recorded *))
+            () (* Only attestation for the wanted proposal are recorded *))
     | "pqc_while_waiting_for_application.v0" -> (
         (* This will never be reached on Mumbai *)
         match (!status.validated, !status.applied) with
@@ -438,22 +438,22 @@ let baker_early_preendorsement_test =
   (* Watch the baker events until the new block is applied and consensus
      operations from their delegates are injected in the node *)
   let status1 = ref (init_status 3) in
-  let preendorsement_t1, preendorsement_u1 = Lwt.task () in
+  let preattestation_t1, preattestation_u1 = Lwt.task () in
   let pqc_t1, pqc_u1 = Lwt.task () in
-  let endorsements_t1, endorsements_u1 = Lwt.task () in
+  let attestations_t1, attestations_u1 = Lwt.task () in
   Baker.on_event
     baker1
-    (baker_on_event ~log:true status1 preendorsement_u1 pqc_u1 endorsements_u1) ;
+    (baker_on_event ~log:true status1 preattestation_u1 pqc_u1 attestations_u1) ;
 
   (* Watch the baker events until the new block is applied and consensus
      operations from their delegates are injected in the node *)
   let status2 = ref (init_status 2) in
-  let preendorsement_t2, preendorsement_u2 = Lwt.task () in
+  let preattestation_t2, preattestation_u2 = Lwt.task () in
   let pqc_t2, pqc_u2 = Lwt.task () in
-  let endorsements_t2, endorsements_u2 = Lwt.task () in
+  let attestations_t2, attestations_u2 = Lwt.task () in
   Baker.on_event
     baker2
-    (baker_on_event status2 preendorsement_u2 pqc_u2 endorsements_u2) ;
+    (baker_on_event status2 preattestation_u2 pqc_u2 attestations_u2) ;
 
   Log.info
     "Continue %s and wait for it to propose a block at level:%d that contains \
@@ -462,74 +462,74 @@ let baker_early_preendorsement_test =
     test_lvl ;
   let* _baker = Baker.continue baker1 in
 
-  let preendorsement_should_be_applied = Protocol.(number protocol > 16) in
+  let preattestation_should_be_applied = Protocol.(number protocol > 16) in
 
-  (* Wait for the pre-endorsements to be included and for the pqc to be reached
+  (* Wait for the pre-attestations to be included and for the pqc to be reached
      before application *)
-  let get_preendorsement_mempool =
+  let get_preattestation_mempool =
     get_mempool
-      ~applied:preendorsement_should_be_applied
-      ~branch_delayed:(not preendorsement_should_be_applied)
+      ~applied:preattestation_should_be_applied
+      ~branch_delayed:(not preattestation_should_be_applied)
   in
-  let* () = if preendorsement_should_be_applied then pqc_t1 else unit
-  and* () = preendorsement_t1 in
-  let* mempool1 = get_preendorsement_mempool client1 in
+  let* () = if preattestation_should_be_applied then pqc_t1 else unit
+  and* () = preattestation_t1 in
+  let* mempool1 = get_preattestation_mempool client1 in
 
-  let* () = if preendorsement_should_be_applied then pqc_t2 else unit
-  and* () = preendorsement_t2 in
-  let* mempool2 = get_preendorsement_mempool client2 in
+  let* () = if preattestation_should_be_applied then pqc_t2 else unit
+  and* () = preattestation_t2 in
+  let* mempool2 = get_preattestation_mempool client2 in
   Log.info
-    "%s have proposed, both baker have pre-endorsed for their delegates%s"
+    "%s have proposed, both baker have pre-attested for their delegates%s"
     (Baker.name baker1)
-    (if preendorsement_should_be_applied then
-     " and the pqc is reached. Ensure that the pre-endorsements are injected \
+    (if preattestation_should_be_applied then
+     " and the pqc is reached. Ensure that the pre-attestations are injected \
       in the nodes as applied operation"
     else
-      ". Ensure that the pre-endorsements are injected in the nodes as \
+      ". Ensure that the pre-attestations are injected in the nodes as \
        branch_delayed operation") ;
 
-  let expected_preendorsements status_a status_b =
+  let expected_preattestations status_a status_b =
     List.of_seq
     @@
-    if preendorsement_should_be_applied then
+    if preattestation_should_be_applied then
       String_set.to_seq
-        (String_set.union !status_a.preendorsements !status_b.preendorsements)
-    else String_set.to_seq !status_a.preendorsements
+        (String_set.union !status_a.preattestations !status_b.preattestations)
+    else String_set.to_seq !status_a.preattestations
   in
   check_ophs_in_mempool
-    ~applied:preendorsement_should_be_applied
-    (expected_preendorsements status1 status2)
+    ~applied:preattestation_should_be_applied
+    (expected_preattestations status1 status2)
     mempool1 ;
   check_ophs_in_mempool
-    ~applied:preendorsement_should_be_applied
-    (expected_preendorsements status2 status1)
+    ~applied:preattestation_should_be_applied
+    (expected_preattestations status2 status1)
     mempool2 ;
 
   Log.info
-    "Wait for the block to be applied by the nodes and wait for them to \
-     endorse it." ;
-  (* Wait for the endorsements to be included *)
-  let* () = endorsements_t1 in
+    "Wait for the block to be applied by the nodes and wait for them to attest \
+     it." ;
+  (* Wait for the attestations to be included *)
+  let* () = attestations_t1 in
   let* mempool1 = get_mempool ~applied:true ~branch_delayed:false client1 in
 
-  let* () = endorsements_t2 in
+  let* () = attestations_t2 in
   let* mempool2 = get_mempool ~applied:true ~branch_delayed:false client2 in
   Log.info
-    "Ensure that the endorsements as well as the pre-endorsements are applied \
+    "Ensure that the attestations as well as the pre-attestations are applied \
      in the mempools" ;
   check_ophs_in_mempool
     ~applied:true
     (List.of_seq
     @@ Seq.append
-         (String_set.to_seq !status1.preendorsements)
-         (String_set.to_seq !status1.endorsements))
+         (String_set.to_seq !status1.preattestations)
+         (String_set.to_seq !status1.attestations))
     mempool1 ;
   check_ophs_in_mempool
     ~applied:true
     (List.of_seq
     @@ Seq.append
-         (String_set.to_seq !status2.preendorsements)
-         (String_set.to_seq !status2.endorsements))
+         (String_set.to_seq !status2.preattestations)
+         (String_set.to_seq !status2.attestations))
     mempool2 ;
 
   Log.info
@@ -546,7 +546,7 @@ let baker_early_preendorsement_test =
     ~error_msg:"Block should contain all the transactions" ;
 
   Log.info
-    "Wait for lvl %d and ensure that it contain all the endorsements"
+    "Wait for lvl %d and ensure that it contain all the attestations"
     next_lvl ;
   let* _ = Node.wait_for_level node2 5 in
   let* consensus_ops =
@@ -560,4 +560,4 @@ let baker_early_preendorsement_test =
     ~error_msg:(sf "Block should contain 5 consensus operations") ;
   unit
 
-let register ~protocols = baker_early_preendorsement_test protocols
+let register ~protocols = baker_early_preattestation_test protocols